Move optimizer model tests to new file
